STY: Move older methods to OMMBV.heritage
