+++ setup.py
@@ -32,8 +32,8 @@ else:
 setup(name='OMMBV',
       version=version,
       license='BSD-3-Clause',
-      packages=['OMMBV', 'OMMBV.satellite', 'OMMBV.trans', 'OMMBV.utils',
-                'OMMBV.vector'],
+      packages=['OMMBV', 'OMMBV.heritage', 'OMMBV.satellite', 'OMMBV.trans',
+                'OMMBV.utils', 'OMMBV.vector'],
       description=' '.join(('Orthogonal geomagnetic vector basis and',
                             'field-line mapping for multipole magnetic',
                             'fields.')),
